fix network data refersh always showing
The snackbar that notifies the user that data has been loaded from
the network is being shown even if there was no prior request
for data refresh.
Added a variable in the MainViewModel that keeps track of pending
network requests and it is used to check if a notification should
be shown to the user.